MODQM-2 marc-json <-> ui-compatible-marc-json converter * Fixed code smells * Added tests
