+++ src/main/java/org/folio/exeptions/EmptyRawRecordException.java
@@ -1,0 +1,7 @@
+package org.folio.exeptions;
+
+public class EmptyRawRecordException extends RuntimeException {
+  public EmptyRawRecordException(){
+    super("RawRecord content should not be empty");
+  }
+}
