+++ src/main/java/org/folio/exeptions/WrongField008LengthException.java
@@ -1,0 +1,7 @@
+package org.folio.exeptions;
+
+public class WrongField008LengthException extends RuntimeException {
+  public WrongField008LengthException() {
+    super("Filed 008 must be 40 characters length");
+  }
+}
