+++ src/test/java/TestSuite.java
@@ -4,6 +4,9 @@ import org.junit.runner.RunWith;
 
 @RunWith(JUnitPlatform.class)
 @SelectClasses({
+  ContentTypeTests.class,
+  Field008SpliterFactoryTests.class,
+  Field008RestoreFactoryTests.class,
   RecordToQuickMarcConverterTest.class,
   QuickMarcToRecordConverterTest.class
 })
